adds syntax for emitter in 0.31.0
- cleans up types
- converts var to const/let
- match order of inputs to model properties for improved readiability
- structure similar to other excalibur projects with src, prettier, etc
