+++ .prettierrc
@@ -1,0 +1,8 @@
+{
+  "printWidth": 140,
+  "semi": true,
+  "singleQuote": true,
+  "tabWidth": 2,
+  "trailingComma": "none",
+  "endOfLine": "auto"
+}
