+++ package.json
@@ -16,11 +16,11 @@
   "author": "Erik Onarheim",
   "license": "BSD-2-Clause",
   "dependencies": {
-    "excalibur": "0.29.3",
+    "excalibur": "0.31.0",
     "knockout": "3.5.1",
-    "typescript": "5.6.2"
+    "typescript": "5.9.3"
   },
   "devDependencies": {
-    "vite": "5.4.8"
+    "vite": "7.1.9"
   }
 }
