+++ src/globals.d.ts
@@ -1,0 +1,9 @@
+export {};
+
+declare global {
+  interface Window {
+    Prism: {
+      highlightAll(): void;
+    };
+  }
+}
